undoing a float precision conversion

+++ template-project/src/Constants.hpp
@@ -41,5 +41,5 @@ namespace Constants {
     static src::Drivers* drivers = src::DoNotUse_getDrivers();
 
     //PI
-    inline constexpr float pi = 3.1415926;
+    inline constexpr double pi = 3.1415926;
 }
